Don't create RTCIceCandidate object
According to spec this is not needed, the RTCIceCandidateInit object
should be just fine. This change makes it easier to adhere to the spec,
because according to the spec the passed candidate may be null.

Also prepended _ to function name in order to provide a wrapper
translating `Nullable RTCIceCandidateInit` to/from `Maybe
RTCIceCandidateInit`.